rft: 拆分 设置-界面设置
