+++ src/MaaWpfGui/Helper/DataHelper.cs
@@ -17,6 +17,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using MaaWpfGui.Constants;
+using MaaWpfGui.ViewModels.UI;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
@@ -30,6 +31,17 @@ namespace MaaWpfGui.Helper
             { "en-us", "YoStarEN" },
             { "ja-jp", "YoStarJP" },
             { "ko-kr", "YoStarKR" },
+        };
+
+        public static readonly Dictionary<string, string> ClientLanguageMapper = new()
+        {
+            { string.Empty, "zh-cn" },
+            { "Official", "zh-cn" },
+            { "Bilibili", "zh-cn" },
+            { "YoStarEN", "en-us" },
+            { "YoStarJP", "ja-jp" },
+            { "YoStarKR", "ko-kr" },
+            { "txwy", "zh-tw" },
         };
 
         // 储存角色信息的字典
@@ -160,7 +172,7 @@ namespace MaaWpfGui.Helper
                 return null;
             }
 
-            language ??= Instances.SettingsViewModel.OperNameLocalization;
+            language ??= SettingsViewModel.GuiSettings.OperNameLocalization;
 
             return language switch
             {
