+++ src/MaaWpfGui/ViewModels/LogItemViewModel.cs
@@ -14,6 +14,7 @@
 using System;
 using MaaWpfGui.Constants;
 using MaaWpfGui.Helper;
+using MaaWpfGui.ViewModels.UI;
 using Stylet;
 
 namespace MaaWpfGui.ViewModels
@@ -32,11 +33,7 @@ namespace MaaWpfGui.ViewModels
         /// <param name="dateFormat">The Date format string</param>
         public LogItemViewModel(string content, string color = UiLogColor.Message, string weight = "Regular", string dateFormat = "MM'-'dd'  'HH':'mm':'ss", bool showTime = true)
         {
-            if (Instances.SettingsViewModel.UseLogItemDateFormat)
-            {
-                dateFormat = Instances.SettingsViewModel.LogItemDateFormatString;
-            }
-
+            dateFormat = SettingsViewModel.GuiSettings.LogItemDateFormatString;
             Time = DateTime.Now.ToString(dateFormat);
             Content = content;
             Color = color;
