+++ src/MaaWpfGui/ViewModels/UI/SettingsViewModel.cs
@@ -71,11 +71,6 @@ namespace MaaWpfGui.ViewModels.UI
         private static extern bool IsIconic(IntPtr hWnd);
 
         /// <summary>
-        /// The Pallas language key.
-        /// </summary>
-        public const string PallasLangKey = "pallas";
-
-        /// <summary>
         /// Gets the visibility of task setting views.
         /// </summary>
         public TaskSettingVisibilityInfo TaskSettingVisibilities { get; } = TaskSettingVisibilityInfo.Current;
@@ -84,6 +79,11 @@ namespace MaaWpfGui.ViewModels.UI
         /// Gets the after action setting.
         /// </summary>
         public PostActionSetting PostActionSetting { get; } = PostActionSetting.Current;
+
+        /// <summary>
+        /// Gets 界面设置model
+        /// </summary>
+        public static GUISettingsUserControlModel GuiSettings { get; } = new();
 
         /// <summary>
         /// Gets 连接设置model
@@ -109,7 +109,7 @@ namespace MaaWpfGui.ViewModels.UI
 
             Init();
 
-            HangoverEnd();
+            GuiSettings.HangoverEnd();
 
             _runningState = RunningState.Instance;
         }
@@ -157,7 +157,7 @@ namespace MaaWpfGui.ViewModels.UI
             InitInfrast();
             InitRoguelike();
             InitConfiguration();
-            InitUiSettings();
+            GuiSettings.InitUiSettings();
             InitConnectConfig();
             InitVersionUpdate();
         }
@@ -236,18 +236,6 @@ namespace MaaWpfGui.ViewModels.UI
             ConfigurationList = configurations;
         }
 
-        private void InitUiSettings()
-        {
-            var languageList = (from pair in LocalizationHelper.SupportedLanguages
-                                where pair.Key != PallasLangKey || Cheers
-                                select new CombinedData { Display = pair.Value, Value = pair.Key })
-               .ToList();
-
-            LanguageList = languageList;
-
-            SwitchDarkMode();
-        }
-
         private void InitConnectConfig()
         {
             var addressListJson = ConfigurationHelper.GetValue(ConfigurationKeys.AddressHistory, string.Empty);
@@ -266,122 +254,6 @@ namespace MaaWpfGui.ViewModels.UI
         }
 
         #endregion Init
-
-        #region EasterEggs
-
-        private bool _cheers = Convert.ToBoolean(ConfigurationHelper.GetValue(ConfigurationKeys.Cheers, bool.FalseString));
-
-        /// <summary>
-        /// Gets or sets a value indicating whether to cheer.
-        /// </summary>
-        public bool Cheers
-        {
-            get => _cheers;
-            set
-            {
-                if (_cheers == value)
-                {
-                    return;
-                }
-
-                SetAndNotify(ref _cheers, value);
-                ConfigurationHelper.SetValue(ConfigurationKeys.Cheers, value.ToString());
-                if (_cheers)
-                {
-                    ConfigurationHelper.SetValue(ConfigurationKeys.Localization, PallasLangKey);
-                }
-            }
-        }
-
-        private bool _hangover = Convert.ToBoolean(ConfigurationHelper.GetValue(ConfigurationKeys.Hangover, bool.FalseString));
-
-        /// <summary>
-        /// Gets or sets a value indicating whether to hangover.
-        /// </summary>
-        public bool Hangover
-        {
-            get => _hangover;
-            set
-            {
-                SetAndNotify(ref _hangover, value);
-                ConfigurationHelper.SetValue(ConfigurationKeys.Hangover, value.ToString());
-            }
-        }
-
-        private string _lastBuyWineTime = ConfigurationHelper.GetGlobalValue(ConfigurationKeys.LastBuyWineTime, DateTime.UtcNow.ToYjDate().AddDays(-1).ToFormattedString());
-
-        public string LastBuyWineTime
-        {
-            get => _lastBuyWineTime;
-            set
-            {
-                SetAndNotify(ref _lastBuyWineTime, value);
-                ConfigurationHelper.SetGlobalValue(ConfigurationKeys.LastBuyWineTime, value);
-            }
-        }
-
-        public void HangoverEnd()
-        {
-            if (!Hangover)
-            {
-                return;
-            }
-
-            Hangover = false;
-            MessageBoxHelper.Show(
-                LocalizationHelper.GetString("Hangover"),
-                LocalizationHelper.GetString("Burping"),
-                iconKey: "HangoverGeometry",
-                iconBrushKey: "PallasBrush");
-            Bootstrapper.ShutdownAndRestartWithoutArgs();
-        }
-
-        public void Sober()
-        {
-            if (!Cheers || Language != PallasLangKey)
-            {
-                return;
-            }
-
-            ConfigurationHelper.SetValue(ConfigurationKeys.Localization, SoberLanguage);
-            Hangover = true;
-            Cheers = false;
-        }
-
-        private string _soberLanguage = ConfigurationHelper.GetValue(ConfigurationKeys.SoberLanguage, LocalizationHelper.DefaultLanguage);
-
-        public string SoberLanguage
-        {
-            get => _soberLanguage;
-            set
-            {
-                SetAndNotify(ref _soberLanguage, value);
-                ConfigurationHelper.SetValue(ConfigurationKeys.SoberLanguage, value);
-            }
-        }
-
-        /// <summary>
-        /// Did you buy wine?
-        /// </summary>
-        /// <returns>The answer.</returns>
-        public bool DidYouBuyWine()
-        {
-            var now = DateTime.UtcNow.ToYjDate();
-            if (now == DateTime.ParseExact(LastBuyWineTime.Replace('-', '/'), "yyyy/MM/dd HH:mm:ss", CultureInfo.InvariantCulture))
-            {
-                return false;
-            }
-
-            if (now.IsAprilFoolsDay())
-            {
-                return true;
-            }
-
-            string[] wineList = ["酒", "liquor", "drink", "wine", "beer", "술", "🍷", "🍸", "🍺", "🍻", "🥃", "🍶"];
-            return wineList.Any(CreditFirstList.Contains);
-        }
-
-        #endregion EasterEggs
 
         #region Remote Control
 
@@ -1565,7 +1437,7 @@ namespace MaaWpfGui.ViewModels.UI
                             continue;
                         }
 
-                        var localizedName = DataHelper.GetLocalizedCharacterName(name, OperNameLocalization);
+                        var localizedName = DataHelper.GetLocalizedCharacterName(name, GuiSettings.OperNameLocalization);
                         if (!string.IsNullOrEmpty(localizedName) && !(_clientType.Contains("YoStar") && DataHelper.GetLocalizedCharacterName(name, "en-us") == DataHelper.GetLocalizedCharacterName(name, "zh-cn")))
                         {
                             roguelikeCoreCharList.Add(localizedName);
@@ -2122,7 +1994,7 @@ namespace MaaWpfGui.ViewModels.UI
             {
                 if (string.IsNullOrEmpty(_reclamationToolToCraft))
                 {
-                    return LocalizationHelper.GetString("ReclamationToolToCraftPlaceholder", _clientLanguageMapper[_clientType]);
+                    return LocalizationHelper.GetString("ReclamationToolToCraftPlaceholder", DataHelper.ClientLanguageMapper[_clientType]);
                 }
 
                 return _reclamationToolToCraft;
@@ -3233,465 +3105,6 @@ namespace MaaWpfGui.ViewModels.UI
 
         #endregion 自动公招设置
 
-        #region 界面设置
-
-        /// <summary>
-        /// Gets or sets the language list.
-        /// </summary>
-        public List<CombinedData> LanguageList { get; set; }
-
-        /// <summary>
-        /// Gets or sets the list of operator name language settings
-        /// </summary>
-        public List<CombinedData> OperNameLanguageModeList { get; set; } =
-            [
-                new() { Display = LocalizationHelper.GetString("OperNameLanguageMAA"), Value = "OperNameLanguageMAA" },
-                new() { Display = LocalizationHelper.GetString("OperNameLanguageClient"), Value = "OperNameLanguageClient" }
-            ];
-
-        /// <summary>
-        /// Gets the list of dark mode.
-        /// </summary>
-        public List<GenericCombinedData<DarkModeType>> DarkModeList { get; } =
-            [
-                new() { Display = LocalizationHelper.GetString("Light"), Value = DarkModeType.Light },
-                new() { Display = LocalizationHelper.GetString("Dark"), Value = DarkModeType.Dark },
-                new() { Display = LocalizationHelper.GetString("SyncWithOs"), Value = DarkModeType.SyncWithOs },
-            ];
-
-        /// <summary>
-        /// Gets the list of inverse clear modes.
-        /// </summary>
-        public List<CombinedData> InverseClearModeList { get; } =
-            [
-                new() { Display = LocalizationHelper.GetString("Clear"), Value = "Clear" },
-                new() { Display = LocalizationHelper.GetString("Inverse"), Value = "Inverse" },
-                new() { Display = LocalizationHelper.GetString("Switchable"), Value = "ClearInverse" },
-            ];
-
-        private bool _useTray = Convert.ToBoolean(ConfigurationHelper.GetValue(ConfigurationKeys.UseTray, bool.TrueString));
-
-        /// <summary>
-        /// Gets or sets a value indicating whether to use tray icon.
-        /// </summary>
-        public bool UseTray
-        {
-            get => _useTray;
-            set
-            {
-                if (!value)
-                {
-                    MinimizeToTray = false;
-                }
-
-                SetAndNotify(ref _useTray, value);
-                ConfigurationHelper.SetValue(ConfigurationKeys.UseTray, value.ToString());
-                Instances.MainWindowManager.SetUseTrayIcon(value);
-            }
-        }
-
-        private bool _minimizeToTray = Convert.ToBoolean(ConfigurationHelper.GetValue(ConfigurationKeys.MinimizeToTray, bool.FalseString));
-
-        /// <summary>
-        /// Gets or sets a value indicating whether to minimize to tray.
-        /// </summary>
-        public bool MinimizeToTray
-        {
-            get => _minimizeToTray;
-            set
-            {
-                SetAndNotify(ref _minimizeToTray, value);
-                ConfigurationHelper.SetValue(ConfigurationKeys.MinimizeToTray, value.ToString());
-                Instances.MainWindowManager.SetMinimizeToTray(value);
-            }
-        }
-
-        private bool _windowTitleScrollable = Convert.ToBoolean(ConfigurationHelper.GetValue(ConfigurationKeys.WindowTitleScrollable, bool.FalseString));
-
-        /// <summary>
-        /// Gets or sets a value indicating whether to make window title scrollable.
-        /// </summary>
-        public bool WindowTitleScrollable
-        {
-            get => _windowTitleScrollable;
-            set
-            {
-                SetAndNotify(ref _windowTitleScrollable, value);
-                ConfigurationHelper.SetValue(ConfigurationKeys.WindowTitleScrollable, value.ToString());
-                var rvm = (RootViewModel)this.Parent;
-                rvm.WindowTitleScrollable = value;
-            }
-        }
-
-        private bool _hideCloseButton = Convert.ToBoolean(ConfigurationHelper.GetValue(ConfigurationKeys.HideCloseButton, bool.FalseString));
-
-        /// <summary>
-        /// Gets or sets a value indicating whether to hide close button.
-        /// </summary>
-        public bool HideCloseButton
-        {
-            get => _hideCloseButton;
-            set
-            {
-                SetAndNotify(ref _hideCloseButton, value);
-                ConfigurationHelper.SetValue(ConfigurationKeys.HideCloseButton, value.ToString());
-                var rvm = (RootViewModel)this.Parent;
-                rvm.ShowCloseButton = !value;
-            }
-        }
-
-        /// <summary>
-        /// Gets or sets a value indicating whether to use notification.
-        /// </summary>
-        public bool UseNotify
-        {
-            get => ConfigFactory.CurrentConfig.GUI.UseNotify;
-            set
-            {
-                ConfigFactory.CurrentConfig.GUI.UseNotify = value;
-                NotifyOfPropertyChange();
-                if (value)
-                {
-                    ToastNotification.ShowDirect("Test test");
-                }
-            }
-        }
-
-        private bool _useLogItemDateFormat = true; // Convert.ToBoolean(ConfigurationHelper.GetValue(ConfigurationKeys.UseLogItemDateFormat, bool.FalseString));
-
-        public bool UseLogItemDateFormat
-        {
-            get => _useLogItemDateFormat;
-            set
-            {
-                SetAndNotify(ref _useLogItemDateFormat, value);
-                ConfigurationHelper.SetValue(ConfigurationKeys.UseLogItemDateFormat, value.ToString());
-            }
-        }
-
-        public List<string> LogItemDateFormatStringList { get; } =
-        [
-            "HH:mm:ss",
-            "MM-dd  HH:mm:ss",
-            "MM/dd  HH:mm:ss",
-            "MM.dd  HH:mm:ss",
-            "dd-MM  HH:mm:ss",
-            "dd/MM  HH:mm:ss",
-            "dd.MM  HH:mm:ss",
-        ];
-
-        private string _logItemDateFormatString = ConfigurationHelper.GetValue(ConfigurationKeys.LogItemDateFormat, "HH:mm:ss");
-
-        public string LogItemDateFormatString
-        {
-            get => _logItemDateFormatString;
-            set
-            {
-                SetAndNotify(ref _logItemDateFormatString, value);
-                ConfigurationHelper.SetValue(ConfigurationKeys.LogItemDateFormat, value);
-            }
-        }
-
-        /// <summary>
-        /// Gets or sets the dark mode.
-        /// </summary>
-        public DarkModeType DarkMode
-        {
-            get => ConfigFactory.CurrentConfig.GUI.DarkMode;
-            set
-            {
-                ConfigFactory.CurrentConfig.GUI.DarkMode = value;
-                NotifyOfPropertyChange();
-                SwitchDarkMode();
-
-                /*
-                AskToRestartToApplySettings();
-                */
-            }
-        }
-
-        public void SwitchDarkMode()
-        {
-            DarkModeType darkModeType = ConfigFactory.CurrentConfig.GUI.DarkMode;
-            switch (darkModeType)
-            {
-                case DarkModeType.Light:
-                    ThemeHelper.SwitchToLightMode();
-                    break;
-
-                case DarkModeType.Dark:
-                    ThemeHelper.SwitchToDarkMode();
-                    break;
-
-                case DarkModeType.SyncWithOs:
-                    ThemeHelper.SwitchToSyncWithOsMode();
-                    break;
-
-                default:
-                    throw new ArgumentOutOfRangeException();
-            }
-        }
-
-        private enum InverseClearType
-        {
-            Clear,
-            Inverse,
-            ClearInverse,
-        }
-
-        private InverseClearType _inverseClearMode =
-            Enum.TryParse(ConfigurationHelper.GetValue(ConfigurationKeys.InverseClearMode, InverseClearType.Clear.ToString()), out InverseClearType temp)
-                ? temp
-                : InverseClearType.Clear;
-
-        /// <summary>
-        /// Gets or sets the inverse clear mode.
-        /// </summary>
-        public string InverseClearMode
-        {
-            get => _inverseClearMode.ToString();
-            set
-            {
-                if (!Enum.TryParse(value, out InverseClearType tempEnumValue))
-                {
-                    return;
-                }
-
-                SetAndNotify(ref _inverseClearMode, tempEnumValue);
-                ConfigurationHelper.SetValue(ConfigurationKeys.InverseClearMode, value);
-                switch (tempEnumValue)
-                {
-                    case InverseClearType.Clear:
-                        Instances.TaskQueueViewModel.InverseMode = false;
-                        Instances.TaskQueueViewModel.ShowInverse = false;
-                        Instances.TaskQueueViewModel.SelectedAllWidth = 90;
-                        break;
-
-                    case InverseClearType.Inverse:
-                        Instances.TaskQueueViewModel.InverseMode = true;
-                        Instances.TaskQueueViewModel.ShowInverse = false;
-                        Instances.TaskQueueViewModel.SelectedAllWidth = 90;
-                        break;
-
-                    case InverseClearType.ClearInverse:
-                        Instances.TaskQueueViewModel.ShowInverse = true;
-                        Instances.TaskQueueViewModel.SelectedAllWidth = TaskQueueViewModel.SelectedAllWidthWhenBoth;
-                        break;
-
-                    default:
-                        throw new ArgumentOutOfRangeException();
-                }
-            }
-        }
-
-        private static readonly Dictionary<string, string> _windowTitleAllShowDict = new()
-        {
-            { LocalizationHelper.GetString("ConfigurationName"), "1" },
-            { LocalizationHelper.GetString("ConnectionPreset"), "2" },
-            { LocalizationHelper.GetString("ConnectionAddress"), "3" },
-            { LocalizationHelper.GetString("ClientType"), "4" },
-        };
-
-        private List<string> _windowTitleAllShowList = [.. _windowTitleAllShowDict.Keys];
-
-        public List<string> WindowTitleAllShowList
-        {
-            get => _windowTitleAllShowList;
-            set => SetAndNotify(ref _windowTitleAllShowList, value);
-        }
-
-        private object[] _windowTitleSelectShowList = ConfigurationHelper.GetValue(ConfigurationKeys.WindowTitleSelectShowList, "1 2 3 4")
-            .Split(' ')
-            .Where(s => _windowTitleAllShowDict.ContainsValue(s.ToString()))
-            .Select(s => _windowTitleAllShowDict.FirstOrDefault(pair => pair.Value == s).Key)
-            .ToArray();
-
-        public object[] WindowTitleSelectShowList
-        {
-            get => _windowTitleSelectShowList;
-            set
-            {
-                SetAndNotify(ref _windowTitleSelectShowList, value);
-                UpdateWindowTitle();
-                var config = string.Join(' ', _windowTitleSelectShowList.Cast<string>().Select(s => _windowTitleAllShowDict[s]));
-                ConfigurationHelper.SetValue(ConfigurationKeys.WindowTitleSelectShowList, config);
-            }
-        }
-
-        private string _language = ConfigurationHelper.GetValue(ConfigurationKeys.Localization, LocalizationHelper.DefaultLanguage);
-
-        /// <summary>
-        /// Gets or sets the language.
-        /// </summary>
-        public string Language
-        {
-            get => _language;
-            set
-            {
-                if (value == _language)
-                {
-                    return;
-                }
-
-                if (_language == PallasLangKey)
-                {
-                    Hangover = true;
-                    Cheers = false;
-                }
-
-                if (value != PallasLangKey)
-                {
-                    SoberLanguage = value;
-                }
-
-                // var backup = _language;
-                ConfigurationHelper.SetValue(ConfigurationKeys.Localization, value);
-
-                var mainWindow = Application.Current.MainWindow;
-
-                if (mainWindow != null)
-                {
-                    mainWindow.Show();
-                    mainWindow.WindowState = mainWindow.WindowState = WindowState.Normal;
-                    mainWindow.Activate();
-                }
-
-                var result = MessageBoxHelper.Show(
-                    FormatText("{0}\n{1}", "LanguageChangedTip"),
-                    FormatText("{0}({1})", "Tip"),
-                    MessageBoxButton.OKCancel,
-                    MessageBoxImage.Question,
-                    ok: FormatText("{0}({1})", "Ok"),
-                    cancel: FormatText("{0}({1})", "ManualRestart"));
-                if (result == MessageBoxResult.OK)
-                {
-                    Bootstrapper.ShutdownAndRestartWithoutArgs();
-                }
-
-                SetAndNotify(ref _language, value);
-
-                return;
-
-                string FormatText(string text, string key)
-                    => string.Format(text, LocalizationHelper.GetString(key, value), LocalizationHelper.GetString(key, _language));
-            }
-        }
-
-        /// <summary>
-        /// Gets the language info.
-        /// </summary>
-        public string LanguageInfo
-        {
-            get
-            {
-                var language = (string)Application.Current.Resources["Language"];
-                return language == "Language" ? language : language + " / Language";
-            }
-        }
-
-        private static readonly Dictionary<string, string> _clientLanguageMapper = new()
-        {
-            { string.Empty, "zh-cn" },
-            { "Official", "zh-cn" },
-            { "Bilibili", "zh-cn" },
-            { "YoStarEN", "en-us" },
-            { "YoStarJP", "ja-jp" },
-            { "YoStarKR", "ko-kr" },
-            { "txwy", "zh-tw" },
-        };
-
-        /// <summary>
-        /// Opername display language, can set force display when it was set as "OperNameLanguageForce.en-us"
-        /// </summary>
-        private string _operNameLanguage = ConfigurationHelper.GetValue(ConfigurationKeys.OperNameLanguage, "OperNameLanguageMAA");
-
-        public string OperNameLanguage
-        {
-            get
-            {
-                if (!_operNameLanguage.Contains('.'))
-                {
-                    return _operNameLanguage;
-                }
-
-                if (_operNameLanguage.Split('.')[0] != "OperNameLanguageForce" || !LocalizationHelper.SupportedLanguages.ContainsKey(_operNameLanguage.Split('.')[1]))
-                {
-                    return _operNameLanguage;
-                }
-
-                OperNameLanguageModeList.Add(new CombinedData { Display = LocalizationHelper.GetString("OperNameLanguageForce"), Value = "OperNameLanguageForce" });
-                return "OperNameLanguageForce";
-            }
-
-            set
-            {
-                if (value == _operNameLanguage.Split('.')[0])
-                {
-                    return;
-                }
-
-                switch (value)
-                {
-                    case "OperNameLanguageClient":
-                        ConfigurationHelper.SetValue(ConfigurationKeys.OperNameLanguage, value);
-                        break;
-
-                    case "OperNameLanguageMAA":
-                    default:
-                        ConfigurationHelper.SetValue(ConfigurationKeys.OperNameLanguage, "OperNameLanguageMAA");
-                        break;
-                }
-
-                var mainWindow = Application.Current.MainWindow;
-
-                if (mainWindow != null)
-                {
-                    mainWindow.Show();
-                    mainWindow.WindowState = mainWindow.WindowState = WindowState.Normal;
-                    mainWindow.Activate();
-                }
-
-                var result = MessageBoxHelper.Show(
-                    LocalizationHelper.GetString("LanguageChangedTip"),
-                    LocalizationHelper.GetString("Tip"),
-                    MessageBoxButton.OKCancel,
-                    MessageBoxImage.Question,
-                    ok: LocalizationHelper.GetString("Ok"),
-                    cancel: LocalizationHelper.GetString("ManualRestart"));
-                if (result == MessageBoxResult.OK)
-                {
-                    Bootstrapper.ShutdownAndRestartWithoutArgs();
-                }
-
-                SetAndNotify(ref _operNameLanguage, value);
-            }
-        }
-
-        public string OperNameLocalization
-        {
-            get
-            {
-                if (_operNameLanguage == "OperNameLanguageClient")
-                {
-                    return _clientLanguageMapper[_clientType];
-                }
-
-                if (!_operNameLanguage.Contains('.'))
-                {
-                    return _language;
-                }
-
-                if (_operNameLanguage.Split('.')[0] == "OperNameLanguageForce" && LocalizationHelper.SupportedLanguages.ContainsKey(_operNameLanguage.Split('.')[1]))
-                {
-                    return _operNameLanguage.Split('.')[1];
-                }
-
-                return _language;
-            }
-        }
-
-        #endregion 界面设置
-
         #region HotKey
 
         /// <summary>
@@ -3912,9 +3325,9 @@ namespace MaaWpfGui.ViewModels.UI
                 prefix += " - ";
             }
 
-            List<string> windowTitleSelectShowList = _windowTitleSelectShowList
-                .Where(x => _windowTitleAllShowDict.ContainsKey(x?.ToString() ?? string.Empty))
-                .Select(x => _windowTitleAllShowDict[x?.ToString() ?? string.Empty]).ToList();
+            List<string> windowTitleSelectShowList = GuiSettings.WindowTitleSelectShowList
+                .Where(x => GUISettingsUserControlModel.WindowTitleAllShowDict.ContainsKey(x?.ToString() ?? string.Empty))
+                .Select(x => GUISettingsUserControlModel.WindowTitleAllShowDict[x?.ToString() ?? string.Empty]).ToList();
 
             string currentConfiguration = string.Empty;
             string connectConfigName = string.Empty;
