+++ src/MaaWpfGui/Views/UserControl/GuideUserControl.xaml
@@ -10,11 +10,11 @@
     xmlns:helper="clr-namespace:MaaWpfGui.Helper"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:s="https://github.com/canton7/Stylet"
-    xmlns:styles="clr-namespace:MaaWpfGui.Styles"
     xmlns:ui="clr-namespace:MaaWpfGui.ViewModels.UI"
     xmlns:userControl="clr-namespace:MaaWpfGui.Views.UserControl"
     xmlns:viewModels="clr-namespace:MaaWpfGui.ViewModels"
     xmlns:vm="clr-namespace:MaaWpfGui"
+    d:Background="White"
     d:DataContext="{d:DesignInstance {x:Type ui:SettingsViewModel}}"
     d:DesignWidth="400"
     mc:Ignorable="d">
@@ -90,7 +90,7 @@
                             Width="240"
                             TransitionMode="{Binding GuideTransitionMode}"
                             Visibility="{c:Binding 'GuideStepIndex less= 0'}">
-                            <hc:UniformSpacingPanel Orientation="Vertical" Spacing="40">
+                            <hc:UniformSpacingPanel DataContext="{Binding GuiSettings}" Orientation="Vertical" Spacing="40">
                                 <ComboBox
                                     hc:InfoElement.HorizontalAlignment="Center"
                                     hc:InfoElement.Title="{Binding LanguageInfo, Mode=OneTime}"
